requirements.txt
flask
flask-cors
requests
python-dotenv
beautifulsoup4
RapidFuzz
openai